SAR-5546 refactored group utr code
