+++ app/config/Module.scala
@@ -91,7 +91,7 @@ class Module extends AbstractModule {
     bind(classOf[EnrolmentsService]).to(classOf[EnrolmentsServiceImpl]).asEagerSingleton()
     bind(classOf[TradingDetailsService]).to(classOf[TradingDetailsServiceImpl]).asEagerSingleton()
     bind(classOf[TestIncorporationService]).to(classOf[TestIncorporationServiceImpl]).asEagerSingleton()
-    bind(classOf[GroupService]).to(classOf[GroupServiceImpl]).asEagerSingleton()
+    bind(classOf[GroupServiceDeprecated]).to(classOf[GroupServiceDeprecatedImpl]).asEagerSingleton()
 
     //controllers
     bind(classOf[CancelRegistrationController]).to(classOf[CancelRegistrationControllerImpl]).asEagerSingleton()
@@ -128,7 +128,6 @@ class Module extends AbstractModule {
     bind(classOf[EmailVerificationController]).to(classOf[EmailVerificationControllerImpl]).asEagerSingleton()
     bind(classOf[GroupReliefController]).to(classOf[GroupReliefControllerImpl]).asEagerSingleton()
     bind(classOf[GroupNameController]).to(classOf[GroupNameControllerImpl]).asEagerSingleton()
-    bind(classOf[GroupUtrController]).to(classOf[GroupUtrControllerImpl]).asEagerSingleton()
     bind(classOf[GroupAddressController]).to(classOf[GroupAddressControllerImpl]).asEagerSingleton()
 
     //test controllers
