+++ test/controllers/GroupUtrControllerSpec.scala
@@ -22,16 +22,14 @@ import controllers.groups.GroupUtrController
 import helpers.SCRSSpec
 import models.{Email, NewAddress, _}
 import org.jsoup.Jsoup
+import org.mockito.Matchers
 import org.mockito.Matchers.any
 import org.mockito.Mockito._
-import org.mockito.{ArgumentMatcher, Matchers}
 import org.scalatestplus.mockito.MockitoSugar
 import play.api.i18n.MessagesApi
-import play.api.libs.json.Json
-import play.api.mvc.{Request, Result, Results}
+import play.api.mvc._
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
-import services.GroupPageEnum
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.test.WithFakeApplication
 
@@ -39,168 +37,235 @@ import scala.concurrent.Future
 
 class GroupUtrControllerSpec extends SCRSSpec with WithFakeApplication with MockitoSugar with AuthBuilder {
 
-  class Setup {
-    implicit val r = FakeRequest()
-    val controller = new GroupUtrController {
-      val groupService = mockGroupService
-      val authConnector = mockAuthConnector
-      val keystoreConnector = mockKeystoreConnector
-      val metaDataService = mockMetaDataService
-      override val compRegConnector = mockCompanyRegistrationConnector
-      implicit val appConfig: FrontendAppConfig = fakeApplication.injector.instanceOf[FrontendAppConfig]
-      override val messagesApi = fakeApplication.injector.instanceOf[MessagesApi]
-      val showFunction = showFunc(_: Groups)(_: Request[_])
-      val submitFunction = submitFunc("1")(_: Groups)(_: Request[_])
-    }
+  implicit val appConfig: FrontendAppConfig = fakeApplication.injector.instanceOf[FrontendAppConfig]
 
-    case class funcMatcher(func: Groups => Future[Result]) extends ArgumentMatcher[Groups => Future[Result]] {
-      override def matches(oarg: scala.Any): Boolean = true
-    }
+  class Setup {
+    implicit val request: FakeRequest[AnyContentAsEmpty.type] = FakeRequest()
+    val controller = new GroupUtrController(
+      mockAuthConnector,
+      mockKeystoreConnector,
+      mockGroupService,
+      mockCompanyRegistrationConnector,
+      fakeApplication.injector.instanceOf[MessagesApi]
+    )
   }
 
-  "The GroupUtrController" should {
-    "redirect whilst the user is un authorised when sending a GET" in new Setup {
-      showWithUnauthorisedUser(controller.show()) {
-        result => {
-          val response = await(result)
-          status(response) shouldBe SEE_OTHER
+  "show" should {
+    "The GroupUtrController" should {
+      "redirect whilst the user is un authorised when sending a GET" in new Setup {
+        showWithUnauthorisedUser(controller.show()) {
+          result => {
+            val response = await(result)
+            status(response) shouldBe SEE_OTHER
+          }
         }
       }
     }
-  }
 
 
-  "Redirect the user to post sign in if the user is authorised but has no registration id in session" in new Setup {
-    mockKeystoreFetchAndGet("registrationID", None)
+    "Redirect the user to post sign in if the user is authorised but has no registration id in session" in new Setup {
+      mockKeystoreFetchAndGet("registrationID", None)
 
-    showWithAuthorisedUser(controller.show()) {
-      result => {
-        val response = await(result)
-        status(response) shouldBe SEE_OTHER
-        response.header.headers("Location") shouldBe "/register-your-company/post-sign-in"
+      showWithAuthorisedUser(controller.show()) {
+        result => {
+          val response = await(result)
+          status(response) shouldBe SEE_OTHER
+          response.header.headers("Location") shouldBe "/register-your-company/post-sign-in"
+        }
+      }
+    }
+
+    "Redirect the user to post sign in if the user is authorised but with incorrect status" in new Setup {
+      mockKeystoreFetchAndGet("registrationID", Some("reg123"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("held", ""))
+
+      when(mockGroupService.retrieveGroups(Matchers.any[String])(Matchers.any[HeaderCarrier]))
+        .thenReturn(Future.successful(None))
+
+      showWithAuthorisedUser(controller.show()) {
+
+        result => {
+          val response = await(result)
+          status(response) shouldBe SEE_OTHER
+          response.header.headers("Location") shouldBe "/register-your-company/post-sign-in"
+        }
+      }
+    }
+
+    "display the page whilst the user is authorised" in new Setup {
+      val testGroups = Groups(groupRelief = true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
+        Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
+        None)
+      mockKeystoreFetchAndGet("registrationID", Some("1"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
+      when(mockCompanyRegistrationConnector.retrieveEmail(any())(any()))
+        .thenReturn(Future.successful(Some(Email("verified@email", "GG", linkSent = true, verified = true, returnLinkEmailSent = true))))
+      when(mockGroupService.retrieveGroups(any())(any()))
+        .thenReturn(Future.successful(Some(testGroups)))
+
+      showWithAuthorisedUser(controller.show) {
+        result =>
+          status(result) shouldBe OK
+      }
+    }
+
+    "display an empty page whilst the user is authorised" in new Setup {
+      val testGroups = Groups(groupRelief = true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
+        Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
+        Some(GroupUTR(None)))
+      mockKeystoreFetchAndGet("registrationID", Some("1"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
+      when(mockCompanyRegistrationConnector.retrieveEmail(any())(any()))
+        .thenReturn(Future.successful(Some(Email("verified@email", "GG", linkSent = true, verified = true, returnLinkEmailSent = true))))
+      when(mockGroupService.retrieveGroups(any())(any()))
+        .thenReturn(Future.successful(Some(testGroups)))
+
+      showWithAuthorisedUser(controller.show) {
+        result =>
+          status(result) shouldBe OK
+          Jsoup.parse(contentAsString(result)).getElementById("groupUtr-noutr").attr("checked") shouldBe "checked"
+      }
+    }
+
+    "display and prepop the page whilst the user is authorised" in new Setup {
+      val testGroups = Groups(groupRelief = true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
+        Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
+        Some(GroupUTR(Some("1234567890"))))
+      mockKeystoreFetchAndGet("registrationID", Some("1"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
+      when(mockCompanyRegistrationConnector.retrieveEmail(any())(any()))
+        .thenReturn(Future.successful(Some(Email("verified@email", "GG", linkSent = true, verified = true, returnLinkEmailSent = true))))
+      when(mockGroupService.retrieveGroups(any())(any()))
+        .thenReturn(Future.successful(Some(testGroups)))
+
+      showWithAuthorisedUser(controller.show) {
+        result =>
+          status(result) shouldBe OK
+          Jsoup.parse(contentAsString(result)).getElementById("groupUtr-utr").attr("checked") shouldBe "checked"
+          Jsoup.parse(contentAsString(result)).getElementById("utr").attr("value") shouldBe "1234567890"
+      }
+    }
+
+    "redirect to the address page if the user skipped it" in new Setup {
+      val testGroups = Groups(groupRelief = true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
+        None,
+        None)
+      mockKeystoreFetchAndGet("registrationID", Some("1"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
+      when(mockCompanyRegistrationConnector.retrieveEmail(any())(any()))
+        .thenReturn(Future.successful(Some(Email("verified@email", "GG", linkSent = true, verified = true, returnLinkEmailSent = true))))
+      when(mockGroupService.retrieveGroups(any())(any()))
+        .thenReturn(Future.successful(Some(testGroups)))
+
+      showWithAuthorisedUser(controller.show) {
+        result =>
+          status(result) shouldBe SEE_OTHER
+          redirectLocation(result) should contain(controllers.groups.routes.GroupAddressController.show().url)
+      }
+    }
+
+    "redirect to the company name if the user skipped it" in new Setup {
+      val testGroups = Groups(groupRelief = true, None,
+        None,
+        None)
+      mockKeystoreFetchAndGet("registrationID", Some("1"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
+      when(mockCompanyRegistrationConnector.retrieveEmail(any())(any()))
+        .thenReturn(Future.successful(Some(Email("verified@email", "GG", linkSent = true, verified = true, returnLinkEmailSent = true))))
+      when(mockGroupService.retrieveGroups(any())(any()))
+        .thenReturn(Future.successful(Some(testGroups)))
+
+      showWithAuthorisedUser(controller.show) {
+        result =>
+          status(result) shouldBe SEE_OTHER
+          redirectLocation(result) should contain(controllers.groups.routes.GroupNameController.show().url)
+      }
+    }
+
+    "redirect to the start of the journey if the user's data is incorrect for the page" in new Setup {
+      val testGroups = Groups(groupRelief = false, None,
+        None,
+        None)
+      mockKeystoreFetchAndGet("registrationID", Some("1"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
+      when(mockCompanyRegistrationConnector.retrieveEmail(any())(any()))
+        .thenReturn(Future.successful(Some(Email("verified@email", "GG", linkSent = true, verified = true, returnLinkEmailSent = true))))
+      when(mockGroupService.retrieveGroups(any())(any()))
+        .thenReturn(Future.successful(Some(testGroups)))
+
+      showWithAuthorisedUser(controller.show) {
+        result =>
+          status(result) shouldBe SEE_OTHER
+          redirectLocation(result) should contain(controllers.reg.routes.SignInOutController.postSignIn().url)
       }
     }
   }
 
-  "Redirect the user to post sign in if the user is authorised but with incorrect status" in new Setup {
-    mockKeystoreFetchAndGet("registrationID", Some("reg123"))
-    CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("held", ""))
+  "submit" should {
+    "return a bad request if the form submitted is incorrect" in new Setup {
+      val testGroups = Groups(groupRelief = true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
+        Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
+        None)
+      mockKeystoreFetchAndGet("registrationID", Some("1"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
+      when(mockCompanyRegistrationConnector.retrieveEmail(any())(any()))
+        .thenReturn(Future.successful(Some(Email("verified@email", "GG", linkSent = true, verified = true, returnLinkEmailSent = true))))
+      when(mockGroupService.retrieveGroups(any())(any()))
+        .thenReturn(Future.successful(Some(testGroups)))
 
-    when(mockGroupService.retrieveGroups(Matchers.any[String])(Matchers.any[HeaderCarrier]))
-      .thenReturn(Future.successful(None))
-
-    showWithAuthorisedUser(controller.show()) {
-
-      result => {
-        val response = await(result)
-        status(response) shouldBe SEE_OTHER
-        response.header.headers("Location") shouldBe "/register-your-company/post-sign-in"
+      submitWithAuthorisedUser(controller.submit(), FakeRequest().withFormUrlEncodedBody(
+        "groupUtr" -> "xxx",
+        "utr" -> "1234567890"
+      )) {
+        result =>
+          status(result) shouldBe BAD_REQUEST
       }
     }
-  }
 
-  "display the page whilst the user is authorised" in new Setup {
-    val testGroups = Groups(true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
-      Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
-      None)
-    mockKeystoreFetchAndGet("registrationID", Some("1"))
-    val mockOfFunc = (g: Groups) => Future.successful(Results.Ok(""))
-    CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
-    when(mockCompanyRegistrationConnector.retrieveEmail(any())(any())).thenReturn(Future.successful(Some(Email("verified@email", "GG", true, true, true))))
-    when(mockGroupService.retrieveGroups(any())(any())).thenReturn(Future.successful(Some(testGroups)))
-    val res: Future[Result] = Future.successful(await(controller.showFunction(testGroups, r)))
-    when(mockGroupService.groupsUserSkippedPage(any[Option[Groups]](), any[GroupPageEnum.Value]())(Matchers.argThat(funcMatcher(mockOfFunc)))).thenReturn(res)
 
-    showWithAuthorisedUser(controller.show) {
-      result =>
-        status(result) shouldBe OK
+    "return a 303 if the user has selected Yes and entered a UTR" in new Setup {
+      val testGroups = Groups(groupRelief = true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
+        Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
+        None)
+      mockKeystoreFetchAndGet("registrationID", Some("1"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
+      when(mockCompanyRegistrationConnector.retrieveEmail(any())(any()))
+        .thenReturn(Future.successful(Some(Email("verified@email", "GG", linkSent = true, verified = true, returnLinkEmailSent = true))))
+      when(mockGroupService.retrieveGroups(any())(any()))
+        .thenReturn(Future.successful(Some(testGroups)))
+      when(mockGroupService.updateGroupUtr(any(), any(), any())(any()))
+        .thenReturn(Future.successful(testGroups.copy(groupUTR = Some(GroupUTR(Some("1234567890"))))))
+
+      submitWithAuthorisedUser(controller.submit(), FakeRequest().withFormUrlEncodedBody(
+        "groupUtr" -> "utr",
+        "utr" -> "1234567890"
+      )) {
+        result =>
+          status(result) shouldBe SEE_OTHER
+          result.header.headers("Location") shouldBe "/register-your-company/psc-handoff"
+      }
     }
-  }
 
-  "display an empty page whilst the user is authorised" in new Setup {
-    val testGroups = Groups(true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
-      Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
-      Some(GroupUTR(None)))
-    mockKeystoreFetchAndGet("registrationID", Some("1"))
-    val mockOfFunc = (g: Groups) => Future.successful(Results.Ok(""))
-    CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
-    when(mockCompanyRegistrationConnector.retrieveEmail(any())(any())).thenReturn(Future.successful(Some(Email("verified@email", "GG", true, true, true))))
-    when(mockGroupService.retrieveGroups(any())(any())).thenReturn(Future.successful(Some(testGroups)))
-    val res: Future[Result] = Future.successful(await(controller.showFunction(testGroups, r)))
-    when(mockGroupService.groupsUserSkippedPage(any[Option[Groups]](), any[GroupPageEnum.Value]())(Matchers.argThat(funcMatcher(mockOfFunc)))).thenReturn(res)
+    "return a 303 if the user has selected No for the UTR" in new Setup {
+      val testGroups = Groups(groupRelief = true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
+        Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
+        None)
+      mockKeystoreFetchAndGet("registrationID", Some("1"))
+      CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
+      when(mockCompanyRegistrationConnector.retrieveEmail(any())(any()))
+        .thenReturn(Future.successful(Some(Email("verified@email", "GG", linkSent = true, verified = true, returnLinkEmailSent = true))))
+      when(mockGroupService.retrieveGroups(any())(any()))
+        .thenReturn(Future.successful(Some(testGroups)))
+      when(mockGroupService.updateGroupUtr(any(), any(), any())(any()))
+        .thenReturn(Future.successful(testGroups))
 
-    showWithAuthorisedUser(controller.show) {
-      result =>
-        status(result) shouldBe OK
-        Jsoup.parse(contentAsString(result)).getElementById("groupUtr-noutr").attr("checked") shouldBe "checked"
-    }
-  }
-
-  "return a bad request if the form submitted is incorrect" in new Setup {
-    val testGroups = Groups(true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
-      Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
-      None)
-    mockKeystoreFetchAndGet("registrationID", Some("1"))
-    val mockOfFunc = (g: Groups) => Future.successful(Results.Ok(""))
-    CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
-    when(mockCompanyRegistrationConnector.retrieveEmail(any())(any())).thenReturn(Future.successful(Some(Email("verified@email", "GG", true, true, true))))
-    when(mockGroupService.retrieveGroups(any())(any())).thenReturn(Future.successful(Some(testGroups)))
-    when(mockGroupService.updateGroupUtr(any(), any(), any())(any())).thenReturn(Future.successful(testGroups))
-    val res: Future[Result] = Future.successful(await(controller.submitFunction(testGroups, FakeRequest().withFormUrlEncodedBody("groupUtr" -> "xxx"))))
-    when(mockGroupService.groupsUserSkippedPage(any[Option[Groups]](), any[GroupPageEnum.Value]())(Matchers.argThat(funcMatcher(mockOfFunc)))).thenReturn(res)
-
-    submitWithAuthorisedUser(controller.submit(), FakeRequest().withFormUrlEncodedBody(
-      "groupUtr" -> "xxx",
-      "utr" -> "1234567890"
-    )) {
-      result =>
-        status(result) shouldBe BAD_REQUEST
-    }
-  }
-
-
-  "return a 303 if the user has selected Yes and entered a UTR" in new Setup {
-    val testGroups = Groups(true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
-      Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
-      None)
-    mockKeystoreFetchAndGet("registrationID", Some("1"))
-    val mockOfFunc = (g: Groups) => Future.successful(Results.Ok(""))
-    CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
-    when(mockCompanyRegistrationConnector.retrieveEmail(any())(any())).thenReturn(Future.successful(Some(Email("verified@email", "GG", true, true, true))))
-    when(mockGroupService.retrieveGroups(any())(any())).thenReturn(Future.successful(Some(testGroups)))
-    when(mockGroupService.updateGroupUtr(any(), any(), any())(any())).thenReturn(Future.successful(testGroups))
-    val res: Future[Result] = Future.successful(await(controller.submitFunction(testGroups, FakeRequest().withFormUrlEncodedBody("groupUtr" -> "utr", "utr" -> "1234567890"))))
-    when(mockGroupService.groupsUserSkippedPage(any[Option[Groups]](), any[GroupPageEnum.Value]())(Matchers.argThat(funcMatcher(mockOfFunc)))).thenReturn(res)
-
-    submitWithAuthorisedUser(controller.submit(), FakeRequest().withFormUrlEncodedBody(
-      "groupUtr" -> "utr",
-      "utr" -> "1234567890"
-    )) {
-      result =>
-        status(result) shouldBe SEE_OTHER
-        result.header.headers("Location") shouldBe "/register-your-company/psc-handoff"
-    }
-  }
-
-  "return a 303 if the user has selected No for the UTR" in new Setup {
-    val testGroups = Groups(true, Some(GroupCompanyName("testGroupCompanyname1", "type")),
-      Some(GroupsAddressAndType("type", NewAddress("l1", "l2", None, None, None, None, None))),
-      None)
-    mockKeystoreFetchAndGet("registrationID", Some("1"))
-    val mockOfFunc = (g: Groups) => Future.successful(Results.Ok(""))
-    CTRegistrationConnectorMocks.retrieveCTRegistration(cTDoc("draft", ""))
-    when(mockCompanyRegistrationConnector.retrieveEmail(any())(any())).thenReturn(Future.successful(Some(Email("verified@email", "GG", true, true, true))))
-    when(mockGroupService.retrieveGroups(any())(any())).thenReturn(Future.successful(Some(testGroups)))
-    when(mockGroupService.updateGroupUtr(any(), any(), any())(any())).thenReturn(Future.successful(testGroups))
-    val res: Future[Result] = Future.successful(await(controller.submitFunction(testGroups, FakeRequest().withFormUrlEncodedBody("groupUtr" -> "noutr"))))
-    when(mockGroupService.groupsUserSkippedPage(any[Option[Groups]](), any[GroupPageEnum.Value]())(Matchers.argThat(funcMatcher(mockOfFunc)))).thenReturn(res)
-
-    submitWithAuthorisedUser(controller.submit(), FakeRequest().withFormUrlEncodedBody(
-      "groupUtr" -> "noutr"
-    )) {
-      result =>
-        status(result) shouldBe SEE_OTHER
-        result.header.headers("Location") shouldBe "/register-your-company/psc-handoff"
+      submitWithAuthorisedUser(controller.submit(), FakeRequest().withFormUrlEncodedBody(
+        "groupUtr" -> "noutr"
+      )) {
+        result =>
+          status(result) shouldBe SEE_OTHER
+          result.header.headers("Location") shouldBe "/register-your-company/psc-handoff"
+      }
     }
   }
 }
