+++ test/mocks/SCRSMocks.scala
@@ -69,6 +69,7 @@ trait SCRSMocks extends CompanyContactDetailsServiceMock
   lazy val mockJweCommon = mock[JweCommon]
   lazy val mockConfiguration = mock[Configuration]
   lazy val mockTimeService = mock[TimeService]
+  lazy val mockGroupServiceDeprecated = mock[GroupServiceDeprecated]
   lazy val mockGroupService = mock[GroupService]
 
 
@@ -111,7 +112,7 @@ trait SCRSMocks extends CompanyContactDetailsServiceMock
     reset(mockAddressLookupService)
     reset(mockConfiguration)
     reset(mockTimeService)
-    reset(mockGroupService)
+    reset(mockGroupServiceDeprecated)
     reset(mockMetricsService)
   }
 }
